Move lit-html utils to module
This extracts `mount` and `withRef` into a `lit-html`-specific utils
module.
